requirements.txt
streamlit
requests
python-dotenv
tqdm
langchain-openai
endee
pypdf
